LayoutTests/http/tests/websocket/tests/hybi/websocket-in-sandboxed-iframe.https.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../../../js-test-resources/js-test.js"></script>
    <script src="../../../cookies/resources/cookie-utilities.js"></script>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFramesAsText();
            testRunner.waitUntilDone();

            internals.setTrackingPreventionEnabled(true);
            testRunner.setStatisticsIsRunningTest(true);
            testRunner.setOnlyAcceptFirstPartyCookies(false);
        }
        var keysCount;
        document.cookie = "jsCookie=value; Path=/; Secure;";

        function runTest()
        {
            shouldBeEqualToString("document.location.host", "127.0.0.1:8443");
            let iframeElement = document.createElement("iframe");
            iframeElement.src = "https://127.0.0.1:8443/websocket/tests/hybi/resources/set-cookie-with-websocket-and-echo.https.html";
            iframeElement.sandbox = "allow-scripts";
            debug("<br>Opening 127.0.0.1 iframe to check its cookies.");

            window.onmessage = data => {
                getCookies().then((cookies) => {
                    keysCount = Object.keys(cookies).length;
                    shouldBe("keysCount", "2");
                    shouldHaveCookieWithValue("jsCookie", "value").then(() => {
                        shouldHaveCookieWithValue("foo", "bar").then(() => testRunner.notifyDone());
                    });
                });
            };

            document.getElementById("output").appendChild(iframeElement);
        }
    </script>
</head>
<body>
<div id="output"></div>
<script>
    description("Tests WebSocket Set-Cookie behavior for sandboxed first-parties cookie.");
    testRunner?.setStatisticsShouldBlockThirdPartyCookies(true, runTest, false, true);
</script>
</body>
</html>
