LayoutTests/http/tests/websocket/tests/hybi/websocket-in-sandboxed-iframe.https-expected.txt

Tests WebSocket Set-Cookie behavior for sandboxed first-parties cookie.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS document.location.host is "127.0.0.1:8443"

Opening 127.0.0.1 iframe to check its cookies.
PASS keysCount is 2
PASS Has cookie "jsCookie" with value value.
PASS Has cookie "foo" with value bar.


--------
Frame: '<!--frame1-->'
--------
PASS document.location.host is "127.0.0.1:8443"

Attempting to set first-party cookie 'foo' through cross-origin WebSocket handshake.
PASS successfullyParsed is true

TEST COMPLETE
PASS Do not have cookie "foo".
PASS document.cookie threw exception SecurityError: The operation is insecure..
PASS Do not have cookie "foo".
PASS Has cookie "foo" with value bar.
